fix bugs, write readme

+++ src/interfaces/ICrossChainFacilitator.sol
@@ -42,7 +42,7 @@ interface ICrossChainFacilitator is IGhoFacilitator {
     // Or we can transfer the GHO that's in the contract, and mint the rest of the amount
     function mintGHOForUSDC(uint256 amount, address to) external;
 
-    function redeemUSDCForGHO(uint64 amount, address to) external;
+    function redeemUSDCForGHO(uint256 amount, address to) external;
 
     function sendGHOCrossChain(uint64 chainId, uint256 amount, address to) external payable returns (bytes32 messageId);
